[clang-tidy] refactor ExprSequence out of use-after-move check
Differential Revision: https://reviews.llvm.org/D27700

git-svn-id: https://llvm.org/svn/llvm-project/clang-tools-extra/trunk@290489 91177308-0d34-0410-b5e6-96231b3b80d8
